requirements.txt
IPython
matplotlib
gym
gym_sokoban
peft
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
numpy
pandas
pybind11
ray>=2.10
tensordict<0.6
transformers
vllm==0.6.3
wandb
gymnasium
gymnasium[toy-text]

pyarrow>=15.0.0
pylatexenc
torchdata
debugpy